feat: 예약이 존재하는 시간 삭제 시도 시 예외 처리 로직 추가

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -6,6 +6,7 @@ import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 
 @Service
@@ -13,9 +14,12 @@ import roomescape.repository.ReservationTimeRepository;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository,
+                                  ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<ReservationTime> findAll() {
@@ -33,6 +37,9 @@ public class ReservationTimeService {
     public void delete(Long id) {
         if (!reservationTimeRepository.existsById(id)) {
             throw new NoSuchElementException("[ERROR] 존재하지 않는 ID입니다.");
+        }
+        if (reservationRepository.existsByTimeId(id)) {
+            throw new IllegalArgumentException("[ERROR] 해당 시간에 예약이 존재합니다.");
         }
         reservationTimeRepository.delete(id);
     }
